Fix multiline code annotations
This now works in slides:

  [:foo] * 10
  # =>

Annotations for a long line of code and/or with a long result can now
follow on the next line so that the code doesn’t get out of the slide.

Why this fixes it? Don’t ask me. RCodeTools is not maintained anymore
and it’s documentation is scarce. Looked at the code to figure out the
problem.